Модуляция исправление
