+++ index.html
@@ -233,6 +233,6 @@
   </template>
   
   <script src="js/functions.js"></script>
-  <script type="module" src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
