+++ js/consts.js
@@ -1,0 +1,29 @@
+const DESCRIPTIONS = [
+  'Красивое фото',
+  'Отличный день для съемки', 
+  'Незабываемые впечатления',
+  'Прекрасный вид',
+  'Момент, который стоит запомнить'
+];
+
+const NAMES = [
+  'Артём',
+  'Мария', 
+  'Дмитрий',
+  'Анна',
+  'Сергей',
+  'Екатерина'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.'
+];
+
+const PHOTOS_COUNT = 25;
+const MAX_COMMENTS = 30;
+
+export { DESCRIPTIONS, NAMES, MESSAGES, PHOTOS_COUNT, MAX_COMMENTS };
